Created feature.pretix package
