--- application/src/main/java/net/furizon/backend/feature/event/action/insert/InsertNewEventAction.java
@@ -1,8 +1,0 @@
-package net.furizon.backend.feature.event.action.insert;
-
-import net.furizon.backend.feature.event.Event;
-import org.jetbrains.annotations.NotNull;
-
-public interface InsertNewEventAction {
-    void invoke(@NotNull Event event);
-}
