+++ application/src/main/java/net/furizon/backend/feature/pretix/event/action/insert/InsertNewEventAction.java
@@ -1,0 +1,8 @@
+package net.furizon.backend.feature.pretix.event.action.insert;
+
+import net.furizon.backend.feature.pretix.event.Event;
+import org.jetbrains.annotations.NotNull;
+
+public interface InsertNewEventAction {
+    void invoke(@NotNull Event event);
+}
