+++ application/src/main/java/net/furizon/backend/feature/pretix/event/action/update/UpdateEventAction.java
@@ -1,0 +1,8 @@
+package net.furizon.backend.feature.pretix.event.action.update;
+
+import net.furizon.backend.feature.pretix.event.Event;
+import org.jetbrains.annotations.NotNull;
+
+public interface UpdateEventAction {
+    void invoke(@NotNull Event event);
+}
