+++ application/src/main/java/net/furizon/backend/feature/pretix/organizer/PretixOrganizer.java
@@ -1,0 +1,18 @@
+package net.furizon.backend.feature.pretix.organizer;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Data;
+import org.jetbrains.annotations.NotNull;
+
+@Data
+public class PretixOrganizer {
+    @NotNull
+    private final String name;
+
+    @NotNull
+    private final String slug;
+
+    @JsonProperty("public_url")
+    @NotNull
+    private final String publicUrl;
+}
