+++ application/src/main/java/net/furizon/backend/feature/pretix/question/PretixQuestion.java
@@ -1,0 +1,16 @@
+package net.furizon.backend.feature.pretix.question;
+
+import lombok.Data;
+import net.furizon.backend.infrastructure.pretix.model.QuestionType;
+import org.jetbrains.annotations.NotNull;
+
+@Data
+public class PretixQuestion {
+    private final int id;
+
+    @NotNull
+    private final QuestionType type;
+
+    @NotNull
+    private final String identifier;
+}
