--- application/src/main/java/net/furizon/backend/feature/product/finder/PretixProductFinder.java
@@ -1,14 +1,0 @@
-package net.furizon.backend.feature.product.finder;
-
-import net.furizon.backend.feature.product.PretixProduct;
-import net.furizon.backend.infrastructure.pretix.dto.PretixPaging;
-import org.jetbrains.annotations.NotNull;
-
-public interface PretixProductFinder {
-    @NotNull
-    PretixPaging<PretixProduct> getPagedProducts(
-            @NotNull String organizer,
-            @NotNull String event,
-            int page
-    );
-}
