--- application/src/main/java/net/furizon/backend/feature/question/PretixQuestion.java
@@ -1,16 +1,0 @@
-package net.furizon.backend.feature.question;
-
-import lombok.Data;
-import net.furizon.backend.infrastructure.pretix.model.QuestionType;
-import org.jetbrains.annotations.NotNull;
-
-@Data
-public class PretixQuestion {
-    private final int id;
-
-    @NotNull
-    private final QuestionType type;
-
-    @NotNull
-    private final String identifier;
-}
