+++ application/src/main/java/net/furizon/backend/infrastructure/pretix/service/CachedPretixInformation.java
@@ -5,12 +5,12 @@ import com.github.benmanes.caffeine.cache.Caffeine;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import net.furizon.backend.feature.event.Event;
-import net.furizon.backend.feature.event.usecase.ReloadEventsUseCase;
-import net.furizon.backend.feature.product.PretixProductResults;
-import net.furizon.backend.feature.product.usecase.ReloadProductsUseCase;
-import net.furizon.backend.feature.question.PretixQuestion;
-import net.furizon.backend.feature.question.usecase.ReloadQuestionsUseCase;
+import net.furizon.backend.feature.pretix.event.Event;
+import net.furizon.backend.feature.pretix.event.usecase.ReloadEventsUseCase;
+import net.furizon.backend.feature.pretix.product.PretixProductResults;
+import net.furizon.backend.feature.pretix.product.usecase.ReloadProductsUseCase;
+import net.furizon.backend.feature.pretix.question.PretixQuestion;
+import net.furizon.backend.feature.pretix.question.usecase.ReloadQuestionsUseCase;
 import net.furizon.backend.infrastructure.pretix.model.CacheItemTypes;
 import net.furizon.backend.infrastructure.pretix.model.ExtraDays;
 import net.furizon.backend.infrastructure.pretix.model.QuestionType;
@@ -21,7 +21,6 @@ import org.jetbrains.annotations.NotNull;
 import org.springframework.data.util.Pair;
 import org.springframework.stereotype.Service;
 
-import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
